pyproject.toml
[tool.poetry]
name = "api"
version = "0.1.0"
description = "Bud backend post migration"
authors = ["Kevin <kevin.souza@weme.nu>"]

[tool.poetry.dependencies]
python = "^3.13"
django = "^5.1.4"
python-dotenv = "^1.0.1"
psycopg2-binary = "^2.9.10"
pyjwt = {extras = ["crypto"], version = "^2.10.1"}
python-dateutil = "^2.9.0.post0"
uuid6 = "^2024.7.10"
drf-yasg = "^1.21.8"
djangorestframework = "^3.15.2"
django-cors-headers = "^4.7.0"
djangorestframework-camel-case = "^1.4.2"


[tool.poetry.group.dev.dependencies]
django-stubs = "^5.1.2"
django-types = "^0.20.0"
pyright = "^1.1.393"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

###############################:weme
#
# LINTERS
#
#

[tool.pyright]
include = ["*"]
exclude = ["**/__pycache__"]

reportIncompatibleVariableOverride = false

[tool.ruff]
# Exclude a variety of commonly ignored directories.
exclude = [
    ".bzr",
    ".direnv",
    ".eggs",
    ".git",
    ".git-rewrite",
    ".hg",
    ".ipynb_checkpoints",
    ".mypy_cache",
    ".nox",
    ".pants.d",
    ".pyenv",
    ".pytest_cache",
    ".pytype",
    ".ruff_cache",
    ".svn",
    ".tox",
    ".venv",
    ".vscode",
    "__pypackages__",
    "_build",
    "buck-out",
    "build",
    "dist",
    "node_modules",
    "site-packages",
    "venv",
]

# Same as Black.
line-length = 88
indent-width = 4

# Assume Python 3.9
target-version = "py39"

[tool.ruff.lint]
# 1. Enable flake8-bugbear (`B`) rules, in addition to the defaults.
select = ["E4", "E7", "E9", "F", "B"]

# 2. Avoid enforcing line-length violations (`E501`)
ignore = ["E501"]

# 3. Avoid trying to fix flake8-bugbear (`B`) violations.
unfixable = ["B"]

# 4. Ignore `E402` (import violations) in all `__init__.py` files, and in selected subdirectories.
[tool.ruff.lint.per-file-ignores]
"__init__.py" = ["E402"]
"**/{tests,docs,tools}/*" = ["E402"]

[tool.ruff.format]
# 5. Use single quotes in `ruff format`.
quote-style = "single"
